% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataFile.R
\name{checkDataFile}
\alias{checkDataFile}
\title{Check the data file}
\usage{
checkDataFile(data)
}
\arguments{
\item{data}{The data file.}
}
\description{
This function checks if the data file has the required format.
}
\keyword{internal}
