% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createbn.R
\name{createbn}
\alias{createbn}
\title{Creates a Bayesian network without arcs.}
\usage{
createbn(nodeNames, emissionOutputPerNode, seed)
}
\arguments{
\item{nodeNames}{The names that the nodes must have.}

\item{emissionOutputPerNode}{The possible outputs per node.}

\item{seed}{The seed for the randomized functions.}
}
\description{
The \code{createbn} function create a Bayesian network without any arcs. The
nodenames supplied as argument are used to determine the amount of nodes and
their names. Furthermore, the possible emissions from the nodes are begin
retrieved from the \code{emissionOutputPerNode} argument.
}
\keyword{internal}
