% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createbnFromData.R
\name{createbnFromData}
\alias{createbnFromData}
\title{Creates a Bayesian network from a datafile}
\usage{
createbnFromData(data, seed, debug = FALSE)
}
\arguments{
\item{data}{The datafile as specified above.}

\item{seed}{Seed.}

\item{debug}{Switches debug mode on/off.}
}
\description{
The \code{createbnFromData} creates a Bayesian network from the datafile that
is supplied. The \code{$x} parameter of the datafile should contain the
actual data, while the $N component should contain the observation lengths.
}
\details{
The possible outputs of a node are based on the unique values in the dataset.
All unique values from a column (a random variable) are converted to possible
outputs of the node in the BN.
}
\keyword{internal}
