% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomNumbers.R
\name{randomNumbers}
\alias{randomNumbers}
\title{Creates a random distribution over numbers.}
\usage{
randomNumbers(amount, sumTo = 1, seed)
}
\arguments{
\item{amount}{The amount of numbers in the vector.}

\item{sumTo}{When all numbers in the returned vector are sumemd, they sum up
to sumTo.}

\item{seed}{A seed when required.}
}
\description{
The \code{randomNumbers} function returns a vector of numbers that are random
within some bandwidth. The count of all the numbers is equal to \code{sumTo},
which defaults to 1.
}
\keyword{internal}
