% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbn.hssm.R
\name{rbn.hssm}
\alias{rbn.hssm}
\title{Random emission from Bayesian network.}
\usage{
\method{rbn}{hssm}(state, model)
}
\arguments{
\item{state}{The state of the model.}

\item{model}{The model.}
}
\value{
A vector with the samples of each node in the Bayesian network.
}
\description{
\code{rbn.hsmm} returns a random sample from the Bayesian network in the
given state for the given model.
}
\keyword{internal}
